+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - `disabled` prop for ActionMenu items
+- Hide clear button on `InputSearch` when disabled
 
 ## [9.124.2] - 2020-07-09