+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `Tab` active color being set as `outset`.
 - Select's label size on regular variation.
 - Accent on `Select` input change.
+- Conditions' operator label color
 
 ### Added